requirements.txt
crewai
crewai-tools
gemini
langchain-gemini
requests
python-dotenv